+++ test/material/light_green/light_green_accents_test.dart
@@ -1,0 +1,14 @@
+import 'package:eo_color/eo_color.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import '../accent_shades.dart';
+import '../accent_swatch_tester.dart';
+
+void main() {
+  group('Light Green Accent Swatch', () {
+    test('Default 4 shades', () {
+      const AccentSwatchTester(LightGreenAccents(), AccentShades.ofLightGreen())
+          .run();
+    });
+  });
+}